+++ lib/redshift_sanitizer/configuration.rb
@@ -1,0 +1,11 @@
+module RedshiftSanitizer
+  class Configuration
+    attr_accessor :delimeter, :replace, :eof
+
+    def initialize
+      @delimeter = "^"
+      @replace = ""
+      @eof = "\n"
+    end
+  end
+end