+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 ## [0.1.0] - 2019-11-02
+
+1. Remove invalid UTF-8 character
+2. Remove delimeter
+3. Remove end_of_field (eof)
+4. Remove nulls
+5. No single start quote